fix: storage folders drop with composer install
When you install using composer the .gitattributes causes the gitignore
files inside the storage folders (and the root) to disappear. If you
then try to commit the project to git it won't include these folders,
nor show you that you can include them.

With this change we renamed the gitignores to gitkeep to ensure these
files remain part of your installation when installing with composer.
They will then also show up in your git to be versioned.